finished project for now

+++ ecommerce/ecommerce/src/main/java/com/project6/ecommerce/domain/CreateProductRequest.java
@@ -3,13 +3,14 @@ package com.project6.ecommerce.domain;
 import com.project6.ecommerce.domain.entity.Category;
 import com.project6.ecommerce.domain.entity.Product.status;
 
+import java.math.BigDecimal;
 import java.util.Set;
 
 public record CreateProductRequest (
         String name,
         String description,
-        double price,
-        int quantity,
+        BigDecimal price,
+        Integer quantity,
         Set<Long> categoryIds,
         String image_url,
         status status
